updated mdl to newest version

+++ include/mi/neuraylib/imdl_discovery_api.h
@@ -1,5 +1,5 @@
 /***************************************************************************************************
- * Copyright (c) 2018-2019, NVIDIA CORPORATION. All rights reserved.
+ * Copyright (c) 2018-2020, NVIDIA CORPORATION. All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -255,7 +255,7 @@ public:
     ///                 the result (see #mi::neuraylib::IMdl_info::Kind). 
     ///                 By default, all kinds are included.
     virtual const IMdl_discovery_result*  discover(
-        Uint32 filter = IMdl_info::DK_ALL) const = 0;
+        Uint32 filter = static_cast<Uint32>(IMdl_info::DK_ALL)) const = 0;
 };
 
 } // namespace neuraylib
